Finished Code for Backend
This code is probably better written than the frontend that I wrote. I
imagine there are "poor coding practices" in it. But you know what? In
the end, IT WORKS. I avoided using switch statements and spaghetti code,
and luckily Java doesn't have much in the way of that anyway. All that's
left now is a UML diagram for it, and the commenting of each component.